[SERVICE-462] Modularize CSS
