+++ res/provider/ui/resolver/css/resolver.css
@@ -1,146 +1,8 @@
+/*
+    TODO: Add fonts to webpack bundling.  JIRA SERVICE-527.  Once completed we can remove this, 
+    and reference to this css in the HTML.
+*/
 @font-face {
     font-family: Akkurat;
     src: url(../assets/Akkurat.ttf);
-}
-
-:root {
-    /*
-        The standard left padding which everything is pushed over by.
-    */
-    --left-padding: 33px;
-}
-
-* {
-    padding: 0px;
-    margin: 0px;
-    -webkit-user-select: none;
-}
-
-::-webkit-scrollbar {
-    width: 10px;
-  }
-
-::-webkit-scrollbar-track {
-    background:  #303038;
-
-  }
-  
-  ::-webkit-scrollbar-thumb {
-    background: #45454c;
-  }
-  
-body {
-    background: #000000;
-    overflow: hidden;
-    font-family: Akkurat;
-}
-
-h1 {
-    color: white;
-    font-size: 20px;
-    font-weight: 400;
-    letter-spacing: 0.15px;
-}
-
-li {
-    list-style: none;
-}
-
-#container {
-    height: 100%;
-    min-height: 100%;
-    display: flex;
-    flex-direction: column;
-}
-
-/*
-    Header
-*/
-
-#header {
-    height: 92px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
-
-#header h1 {
-    padding-left: var(--left-padding);
-}
-
-#header #exit {
-    height: 20px;
-    width: 20px;
-    margin-right: 30px;
-    background-image: url("../assets/exit.png");
-    background-position: 50%;
-    background-size: 9px 9px;
-    background-repeat: no-repeat;
-}
-
-#header #exit:hover {
-    background-color: #3e3d51;
-    border-radius: 10px;
-}
-
-/*
-    App list
-*/
-
-#app-list {
-    color: white;
-    background: #0a0a0a;
-    display: flex;
-    flex: 1;
-	flex-direction: column;
-}
-
-#app-list p {
-    background: #303038;
-    color: #5a5a63;
-    height: 25px;
-    padding-left: var(--left-padding);
-    font-size: 8px;
-    font-weight: 700;
-    letter-spacing: 1px;
-    line-height: 19px;
-    text-transform: uppercase;
-    display: flex;
-    align-items: center;
-}
-
-#app-list ul {
-    flex: 1;
-    overflow-y: auto;
-}
-
-#app-list li {
-    height: 85px;
-    padding-left: var(--left-padding);
-    display: flex;
-    align-items: center;
-    border-bottom: 1px solid #333333;
-}
-
-#app-list li:last-child {
-    border-bottom: none;
-}
-
-#app-list li:hover {
-    background: #252637;
-}
-
-#app-list li .icon {
-    width: 56px;
-    height: 56px;
-    border-radius: 6px;
-}
-
-#app-list li h1 {
-    margin: 0 18px 0 18px;
-    overflow: hidden;
-    display: -webkit-box;
-    text-overflow: ellipsis;
-    -webkit-box-orient: vertical;
-    -webkit-line-clamp: 2; 
 }
