+++ src/provider/view/components/AppCard.css
@@ -1,0 +1,30 @@
+.app-card {
+    height: 85px;
+    padding-left: var(--left-padding);
+    display: flex;
+    align-items: center;
+    border-bottom: 1px solid #333333;
+}
+
+.app-card:last-child {
+    border-bottom: none;
+}
+
+.app-card:hover {
+    background: #252637;
+}
+
+.app-card>.icon {
+    width: 56px;
+    height: 56px;
+    border-radius: 6px;
+}
+
+.app-card>h1 {
+    margin: 0 18px 0 18px;
+    overflow: hidden;
+    display: -webkit-box;
+    text-overflow: ellipsis;
+    -webkit-box-orient: vertical;
+    -webkit-line-clamp: 2; 
+}
