+++ src/provider/view/components/AppCard.tsx
@@ -2,6 +2,8 @@ import * as React from 'react';
 
 import {Application} from '../../../client/directory';
 
+import './AppCard.css';
+
 interface AppCardProps {
     app: Application;
     openHandler: (app: Application) => void;
@@ -16,7 +18,7 @@ export function AppCard(props: AppCardProps): React.ReactElement {
         openHandler(app);
     };
     return (
-        <li onClick={clickHandler}>
+        <li className="app-card" onClick={clickHandler}>
             {(app.icons && app.icons.length > 0) && <img className="icon" src={app.icons![0].icon} />}
             <h1>{app.title}</h1>
         </li>
